+++ mail/bugReport-text.php
@@ -1,0 +1,14 @@
+<?php
+
+/* @var $this yii\web\View */
+/* @var $user \wdmg\users\models\Users */
+
+?>
+Butterfly.CMS
+===
+
+<?= Yii::t('app/modules/admin', 'Hi!'); ?>
+
+Name:</b> <?= $name ?>
+E-mail:</b> <?= $email ?>
+Message:</b> <?= $message ?>