+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 ## 1.1.7 (2019-09-07)
+ * Added bug report functionality
  * Fix check updates for modules. Add method getAppVersion() for layout. Update dependencies.
  
 ## 1.1.6 (2019-09-05)